# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/library": {
      "get": {
        "operationId": "library_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "PLANNED",
                  "IN_PROGRESS",
                  "COMPLETED",
                  "ON_HOLD",
                  "DROPPED"
                ]
              }
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "GAME",
                  "MOVIE",
                  "TV_SHOW",
                  "BOOK",
                  "ANIME"
                ]
              }
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "recent",
              "type": "string",
              "enum": [
                "name",
                "rating",
                "recent"
              ]
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "library"
        ]
      },
      "post": {
        "operationId": "library_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "library"
        ]
      }
    },
    "/library/by-title/{titleId}": {
      "get": {
        "operationId": "library_findByTitle",
        "parameters": [
          {
            "name": "titleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Есть ли тайтл в библиотеке — для кнопки на карточке и в расширении",
        "tags": [
          "library"
        ]
      }
    },
    "/library/{id}": {
      "patch": {
        "operationId": "library_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "library"
        ]
      },
      "delete": {
        "operationId": "library_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "library"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "notification_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isUnread",
            "required": false,
            "in": "query",
            "description": "Только непрочитанные — для колокольчика",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "notification_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "patch": {
        "operationId": "notification_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "delete": {
        "operationId": "notification_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/reviews/title/{slug}": {
      "get": {
        "operationId": "review_findByTitle",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "rating",
                "newest",
                "ratingAsc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          }
        },
        "summary": "Отзывы на тайтле — открыты без авторизации",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/my": {
      "get": {
        "operationId": "review_findMy",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "rating",
                "newest",
                "ratingAsc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyReviewListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews": {
      "post": {
        "operationId": "review_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "patch": {
        "operationId": "review_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      },
      "delete": {
        "operationId": "review_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "RED Marathōn API",
    "description": "One library for movies, TV shows, anime, books and games",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TitleListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "coverUrl",
          "releaseDate",
          "rating",
          "ratingCount"
        ]
      },
...
      "LibraryEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD",
              "DROPPED"
            ],
            "type": "string"
          },
          "progress": {
            "type": "number",
            "nullable": true
          },
          "progressUnit": {
            "nullable": true,
            "enum": [
              "EPISODE",
              "CHAPTER",
              "PAGE",
              "HOUR",
              "PERCENT"
            ],
            "type": "string"
          },
          "rating": {
            "type": "number",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "title": {
            "$ref": "#/components/schemas/TitleListItemResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "progress",
          "progressUnit",
          "rating",
          "note",
          "isFavorite",
          "startedAt",
          "finishedAt",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
      "LibraryListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryEntryResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
      "CreateLibraryEntryDto": {
        "type": "object",
        "properties": {
          "titleId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD",
              "DROPPED"
            ]
          },
          "progress": {
            "type": "number",
            "minimum": 0
          },
          "progressUnit": {
            "type": "string",
            "enum": [
              "EPISODE",
              "CHAPTER",
              "PAGE",
              "HOUR",
              "PERCENT"
            ]
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          },
          "isFavorite": {
            "type": "boolean"
          },
          "startedAt": {
            "type": "string",
            "description": "Если не прислать — проставится автоматически по статусу"
          },
          "finishedAt": {
            "type": "string"
          }
        },
        "required": [
          "titleId"
        ]
      },
      "UpdateLibraryEntryDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD",
              "DROPPED"
            ],
            "type": "string"
          },
          "progress": {
            "type": "number",
            "minimum": 0
          },
          "progressUnit": {
            "enum": [
              "EPISODE",
              "CHAPTER",
              "PAGE",
              "HOUR",
              "PERCENT"
            ],
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          },
          "isFavorite": {
            "type": "boolean"
          },
          "startedAt": {
            "type": "string",
            "description": "Если не прислать — проставится автоматически по статусу"
          },
          "finishedAt": {
            "type": "string"
          }
        }
      },
...
      "NotificationActorResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "SYSTEM",
              "COLLECTION_SHARED",
              "REVIEW_REPLY",
              "SUBSCRIPTION",
              "RECOMMENDATION"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "entityType": {
            "nullable": true,
            "description": "Куда вести по клику: тип объекта и его id",
            "enum": [
              "COLLECTION",
              "TITLE",
              "REVIEW"
            ],
            "type": "string"
          },
          "entityId": {
            "type": "string",
            "nullable": true
          },
          "actor": {
            "nullable": true,
            "description": "Кто спровоцировал событие, если это не системное уведомление",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationActorResponse"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "message",
          "isRead",
          "readAt",
          "entityType",
          "entityId",
          "actor",
          "createdAt"
        ]
      },
      "NotificationListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          },
          "unreadCount": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "isHasMore",
          "unreadCount"
        ]
      },
      "ReviewAuthorResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/ReviewAuthorResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rating",
          "text",
          "isPublic",
          "author",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReviewListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
      "MyReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "author": {
            "$ref": "#/components/schemas/ReviewAuthorResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/TitleListItemResponse"
          }
        },
        "required": [
          "id",
          "rating",
          "text",
          "isPublic",
          "author",
          "createdAt",
          "updatedAt",
          "title"
        ]
      },
      "MyReviewListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyReviewResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "titleId": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "number",
            "description": "Оценки везде десятибалльные, клиент показывает их по своей шкале",
            "minimum": 1,
            "maximum": 10
          },
          "text": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "titleId",
          "rating"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "description": "Оценки везде десятибалльные, клиент показывает их по своей шкале",
            "minimum": 1,
            "maximum": 10
          },
          "text": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "isPublic": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}